SockJS bug fix.
Ensure that when in livedata_server.js we send version negotiation error and
then close the socket, that the close actually gets delivered. In 0.3.4, this
often caused the livedata tests to spin as the client kept not quite getting the
error and reconnecting.

(Strangely, 0.3.5 on its own seems to fix this. I'm not quite sure why.)